+++ lib/modules/core/core_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: core
-// Updated at: 2026-06-05 17:29:40
-// Commit: refactor(core): decouple repository interface from concrete core service
+// Updated at: 2026-06-05 21:07:45
+// Commit: feat(core): add bi-directional event bus handler for core
 
 class coreComponentState {
-  final String commitRef = '1779';
-  final String lastMessage = 'refactor(core): decouple repository interface from concrete core service';
-  final DateTime timestamp = DateTime.parse('2026-06-05T17:29:40+05:00');
+  final String commitRef = '1789';
+  final String lastMessage = 'feat(core): add bi-directional event bus handler for core';
+  final DateTime timestamp = DateTime.parse('2026-06-05T21:07:45+05:00');
 }